The mixins are dead, long live the mixins!
Complete refactoring of all mixins to be "Real Mixins".
The update inspired by article: http://justinfagnani.com/2015/12/21/real-mixins-with-javascript-classes/

The idea is that every mixin is a generic class where the
  template class is a superclass.
Something like (C# style):

  class MyMixin<T> : T
    where T : class // with baseMethod()

    myProperty;
    override baseMethod() {
       super();
       // my code
    }

This indeed seems to be the most effecient way of mixin with
   abstract classes. It gives ability to override super methods
   and maintain easy to debug prototype inheritance chain.

With this, we can give up flight advice. Bye-bye.