Fix binding in mixins
Studies show that using manual binding in ErrorHandling and
ErrorHandled mixins did not work and context wasn’t preserved.
Also it seems there was no reason in a manual binding, since
we do not try to access bound handler in the constructor.

Add few more checks if the target for mutation is disposed.