Show error notification in some cases
We want to show normal red error notification if
syncing activity failed using same instance of the
collection. Now all those notifications are muted in
favor of showing .error service row and it’s confusing UX
because this row is the last one after existing thirty. It is
hard to notice it.