Make all dependencies peer
This is to underline that `gringotts` is distributed as raw
source code. So the parent project should make sure that
all the NPM dependencies are reused. This is important for unit
testing. Where if we stub somthing in `Backbone.View.prototype`,
we want to make sure this stub affects `gringotts` code.